+++ cpp/slra_common.cpp
@@ -14,7 +14,7 @@ Exception::Exception( const char *format, ... ) {
 }
 
 Structure *createMosaicStructure( gsl_vector * ml,  gsl_vector *nk, 
-               gsl_vector * wk, int np_comp ) {
+               gsl_vector * wk, size_t np_comp ) {
   if (wk == NULL || wk->size == ml->size * nk->size || wk->size == ml->size) { 
     return new MosaicHStructure(ml, nk, wk);
   } 
@@ -29,12 +29,12 @@ void OptimizationOptions::str2Method( const char *str )  {
        sm_codes_lm[] = "ls", sm_codes_qn[] = "b2pf", sm_codes_nm[] = "n2r";
   char *submeth_codes[] = { sm_codes_lm, sm_codes_qn, sm_codes_nm };
 
-  int submeth_codes_max[] = { 
+  size_t submeth_codes_max[] = { 
     sizeof(sm_codes_lm) / sizeof(sm_codes_lm[0]) - 1, 
     sizeof(sm_codes_qn) / sizeof(sm_codes_qn[0]) - 1, 
     sizeof(sm_codes_nm) / sizeof(sm_codes_nm[0]) - 1
   };
-  int meth_code_max = sizeof(submeth_codes_max) / sizeof(submeth_codes_max[0]);
+  size_t meth_code_max = sizeof(submeth_codes_max) / sizeof(submeth_codes_max[0]);
   int i;
   
   if (str[0] == 0) {
@@ -66,8 +66,8 @@ void OptimizationOptions::str2Method( const char *str )  {
 /* gsl_matrix_vectorize: vectorize column-wise a gsl_matrix */
 void gsl_matrix_vectorize(double* v, const gsl_matrix* m)
 {
-  for (int j = 0; j < m->size2; j++) {
-    for (int i = 0; i < m->size1; i++) {
+  for (size_t j = 0; j < m->size2; j++) {
+    for (size_t i = 0; i < m->size1; i++) {
       v[i+j*m->size1] = gsl_matrix_get(m,i,j); 
     }
   }
@@ -78,8 +78,8 @@ void gsl_matrix_vectorize(double* v, const gsl_matrix* m)
 
 void gsl_matrix_vec_inv(gsl_matrix* m, const double* v)
 {
-  for (int i = 0; i < m->size1; i++) {
-    for (int j = 0; j < m->size2; j++) {
+  for (size_t i = 0; i < m->size1; i++) {
+    for (size_t j = 0; j < m->size2; j++) {
       gsl_matrix_set(m, i, j, v[i + j * m->size1]);
     }
   }
@@ -88,7 +88,7 @@ void gsl_matrix_vec_inv(gsl_matrix* m, const double* v)
 /* print matrix */
 void print_mat(const gsl_matrix* m)
 {
-  int i, j;
+  size_t i, j;
 
   PRINTF("\n");
   for (i = 0; i < m->size1; i++) {
@@ -103,7 +103,7 @@ void print_mat(const gsl_matrix* m)
 /* print matrix */
 void print_mat_tr(const gsl_matrix* m)
 {
-  int i, j;
+  size_t i, j;
 
   PRINTF("\n");
   for (j = 0; j < m->size2; j++) {
@@ -117,9 +117,9 @@ void print_mat_tr(const gsl_matrix* m)
 
 
 /* print_arr: print array */
-void print_arr(const double* a, int n)
+void print_arr(const double* a, size_t n)
 {
-  int i;
+  size_t i;
 
   PRINTF("\n");
   for (i = 0; i < n; i++)
@@ -129,7 +129,7 @@ void print_arr(const double* a, int n)
 
 void print_vec(const gsl_vector* a)
 {
-  int i;
+  size_t i;
 
   PRINTF("\n");
   for (i = 0; i < a->size; i++)
@@ -170,23 +170,23 @@ int read_vec_uint( gsl_vector_uint *a, const char * filename ) {
   return 1;
 }
 
-int compute_np( gsl_vector* ml, gsl_vector *nk ) {
-  int np = 0;
-  int i;
+size_t compute_np( gsl_vector* ml, gsl_vector *nk ) {
+  size_t np = 0;
+  size_t i;
   
   for (i = 0; i < ml->size; i++) {
-    np += ((int)gsl_vector_get(ml, i) - 1) * nk->size; 
+    np += ((size_t)gsl_vector_get(ml, i) - 1) * nk->size; 
   }
   for (i = 0; i < nk->size; i++) {
-    np += ((int)gsl_vector_get(nk, i)) * ml->size; 
+    np += ((size_t)gsl_vector_get(nk, i)) * ml->size; 
   }
   return np;
 }
 
 
-int compute_n( gsl_vector* ml, int np ) {
-  for (int i = 0; i < ml->size; i++) {
-    np -= ((int)gsl_vector_get(ml, i) - 1);
+size_t compute_n( gsl_vector* ml, size_t np ) {
+  for (size_t i = 0; i < ml->size; i++) {
+    np -= ((size_t)gsl_vector_get(ml, i) - 1);
   }
 
   if (np <= 0 || (np % ml->size != 0)) {
@@ -197,7 +197,7 @@ int compute_n( gsl_vector* ml, int np ) {
 }
 
 void Cholesky::multInvCholeskyTransMatrix( gsl_matrix * yr_matr, int trans ) { 
-  for (int i = 0; i < yr_matr->size1; i++) {
+  for (size_t i = 0; i < yr_matr->size1; i++) {
     gsl_vector_view row = gsl_matrix_row(yr_matr, i);
     multInvCholeskyVector(&row.vector, trans);
   }
@@ -228,9 +228,9 @@ char *strncpy0( char *dest, const char * src, size_t buf_len ) {
   return res;
 }
 
-void id_kron_a( const gsl_matrix *A, int d,  gsl_matrix *IkronA ) {
+void id_kron_a( const gsl_matrix *A, size_t d,  gsl_matrix *IkronA ) {
   gsl_matrix_set_zero(IkronA);
-  for (int  k = 0; k < d; k++) {
+  for (size_t  k = 0; k < d; k++) {
     gsl_matrix subA = gsl_matrix_submatrix(IkronA, k * A->size1, k * A->size2, 
                            A->size1, A->size2).matrix;
     gsl_matrix_memcpy(&subA, A);
@@ -296,7 +296,7 @@ void Log::setMaxLevel( Level maxLevel ) {
   
 void Log::str2DispLevel( const char *str )  {
   char *str_disp[] = { "off", "final", "notify", "iter" };
-  int i;
+  size_t i;
   
   
   for (i = 0; i < sizeof(str_disp) / sizeof(str_disp[0]); i++) {
